[Fire Mage] Merged Combustion & Converted to TypeScript
I merged the Combustion modules back together and tried my hand at converting the module to TypeScript.

If someone can take a look at this and make sure i didnt screw anything up given that this is my first attempt at TypeScript.
