+++ src/parser/mage/fire/CHANGELOG.js
@@ -8,6 +8,7 @@ import ItemLink from 'common/ItemLink';
 import { change, date } from 'common/changelog';
 
 export default [
+  change(date(2019, 12, 16), <>Merged <SpellLink id={SPELLS.COMBUSTION.id} /> modules together and converted module to Typescript.</>,[Sharrq]),
   change(date(2019, 11, 14), <>Added analysis for using <SpellLink id={SPELLS.METEOR_TALENT.id} /> during <SpellLink id={SPELLS.COMBUSTION.id} />. Also adjusted Meteor Efficiency target to account for holding it for Combustion.</>,[Sharrq]),
   change(date(2019, 10, 9), <>Added Statistic Boxes for <SpellLink id={SPELLS.SEARING_TOUCH_TALENT.id} /> and <SpellLink id={SPELLS.KINDLING_TALENT.id} /> and fixed the <SpellLink id={SPELLS.RUNE_OF_POWER_TALENT.id} /> statistic box.</>,[Sharrq]),
   change(date(2019, 10, 9), <>Combined the statistic boxes for <SpellLink id={SPELLS.HOT_STREAK.id} /> and Hot Streak Pre-Casts.</>,[Sharrq]),
